hitched2/requirements.txt

Django==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7
django-extensions==3.6.4
django-spa==0.2.0
